Split core logic into control and gui
